% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_basel.R
\name{add_parameter_metadata}
\alias{add_parameter_metadata}
\title{Helper function: add parameter metadata}
\usage{
add_parameter_metadata(
  df,
  meta_parameter_path = package_file("shiny/basel/data/metadata/meta_parameter.csv")
)
}
\arguments{
\item{df}{data frame containing at least a column "ParameterCode"}

\item{meta_parameter_path}{Define path of "meta_parameter.csv" to be imported
(default: sema.pilot:::package_file("shiny/basel/data/metadata/meta_parameter.csv"))}
}
\value{
returns input data frame with joined metadata (parameter codes/ methods
not included in meta_parameter file will not be imported!!!!)
}
\description{
Helper function: add parameter metadata
}
