% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_export_fst_general.R
\name{aggregate_export_fst}
\alias{aggregate_export_fst}
\title{Aggregate and Export to FST Format}
\usage{
aggregate_export_fst(
  year_month_start = NULL,
  year_month_end = format(Sys.Date(), "\%Y-\%m"),
  compression = 100,
  FUN_get_monthly_data = NULL,
  FUN_import = NULL,
  FUN_calculate_ops = NULL,
  prefix = NULL,
  calculate_ops_on_level = c("raw", "aggregated")[1L],
  siteData_raw_list = NULL,
  export_dir_path = NULL
)
}
\arguments{
\item{year_month_start}{start year month in yyyy-mm format, e.g. "2017-06"}

\item{year_month_end}{end year month (default: current month)}

\item{compression}{(default: 100)}

\item{FUN_get_monthly_data}{function to be called to determine
"raw_data_file_paths"}

\item{FUN_import}{function to be called to read data into "siteData_raw_list"}

\item{FUN_calculate_ops}{function to be called to calculate operational
parameters. If \code{NULL} (the default), no operational parameters are
calculated}

\item{prefix}{site-specific prefix to be used as a sub folder name in the
export directory}

\item{calculate_ops_on_level}{one of "raw", "aggregated". Determines whether
to calculate operational parameters based on the raw or on the aggregated
values. Not used if \code{FUN_calculate_ops} is \code{NULL}}

\item{siteData_raw_list}{existing raw data. If given (not \code{NULL}), no
monthly periods are calculated and no data are imported. The given data are
treated as one and only (monthly?) period.}

\item{export_dir_path}{path to the export directory. Only required if raw
data are given in \code{siteData_raw}.}
}
\value{
exports data for each month into subfolder: /data/fst/year-month
}
\description{
Aggregate and Export to FST Format
}
