% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_export_fst.R, R/export_c.R
\name{aggregate_export_fst_berlin_f}
\alias{aggregate_export_fst_berlin_f}
\title{Berlin-Friedrichshagen: aggregate and export to fst}
\usage{
aggregate_export_fst_berlin_f(
  year_month_start = "2019-11",
  year_month_end = format(Sys.Date(), "\%Y-\%m"),
  compression = 100
)

aggregate_export_fst_berlin_f(
  year_month_start = "2019-11",
  year_month_end = format(Sys.Date(), "\%Y-\%m"),
  compression = 100
)
}
\arguments{
\item{year_month_start}{start year month (default: '2019-11')}

\item{year_month_end}{end year month (default: current month)}

\item{compression}{(default: 100)}
}
\value{
exports data for each month into subfolder: /data/fst/year-month

exports data for each month into subfolder: /data/fst/year-month
}
\description{
Berlin-Friedrichshagen: aggregate and export to fst

Berlin-Friedrichshagen: aggregate and export to fst
}
