% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_export_fst.R
\name{aggregate_export_fst_berlin_s}
\alias{aggregate_export_fst_berlin_s}
\title{Berlin-Schoenerlinde: aggregate and export to fst}
\usage{
aggregate_export_fst_berlin_s(
  year_month_start = "2017-04",
  year_month_end = format(Sys.Date(), "\%Y-\%m"),
  compression = 100
)
}
\arguments{
\item{year_month_start}{start year month (default: '2017-04')}

\item{year_month_end}{end year month (default: current month)}

\item{compression}{(default: 100)}
}
\value{
exports data for each month into subfolder: /data/fst/year-month
}
\description{
Berlin-Schoenerlinde: aggregate and export to fst
}
