% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_export_fst.R
\name{aggregate_export_fst_mbr4}
\alias{aggregate_export_fst_mbr4}
\title{MBR4.0: aggregate and export to fst}
\usage{
aggregate_export_fst_mbr4(
  mbr4_data_tidy = tidy_mbr4_data(read_mbr4()),
  compression = 100
)
}
\arguments{
\item{mbr4_data_tidy}{tidy MBR4 data as retrieved by \link{tidy_mbr4_data},
(default: kwb.pilot::tidy_mbr4_data(kwb.pilot::read_mbr4()))}

\item{compression}{(default: 100)}
}
\value{
exports data for each month into subfolder: /data/fst/year-month
}
\description{
MBR4.0: aggregate and export to fst
}
