% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_operational_parameters_berlin_s.R
\name{calculate_operational_parameters_berlin_s}
\alias{calculate_operational_parameters_berlin_s}
\title{Calculate operational parameters for Berlin-Schoenerlinde}
\usage{
calculate_operational_parameters_berlin_s(
  df,
  calc_list = get_calc_info_berlin_s(),
  calc_list_name = get_calc_info_berlin_s("name"),
  calc_list_unit = get_calc_info_berlin_s("unit"),
  calc_paras = get_calc_info_berlin_s("paras")
)
}
\arguments{
\item{df}{a data frame as retrieved by read_wedeco_data()}

\item{calc_list}{list with calculation operations to be carried out
(default: list(deltaSAK  = "(1-SCAN_SAK_Ablauf/SCAN_SAK_Zulauf)*100",
Ozoneintrag = "(C_O3_Zugas - C_O3_Abgas)*Q_Gas/Q_Ozonanlage"))}

\item{calc_list_name}{full names of parameters to be used for plotting for each
calculation specified wit 'calc_list'. default: c('delta SAK', 'Ozoneintrag')}

\item{calc_list_unit}{units of parameters to be used for plotting for each
calculation specified wit 'calc_list'. default: c("percent", "mg-O3/L")}

\item{calc_paras}{a vector with parameter codes used for performing calculations
defined in 'calc_list' (default: c("SCAN_SAK_Ablauf", "SCAN_SAK_Zulauf",
"C_O3_Zugas", "C_O3_Abgas", "Q_Gas", "Q_Ozonanlage"))}
}
\value{
dataframe with calculated operational parameters
}
\description{
Calculate operational parameters for Berlin-Schoenerlinde
}
\examples{
\dontrun{
raw_list <- read_wedeco_data()
myDat <- calculate_operational_parameters_berlin_s(df = raw_list)
}

}
