% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monthly_data_from_calenderweeks.R
\name{calenderweek_from_dates}
\alias{calenderweek_from_dates}
\title{Helper function: get calender weeks for time period}
\usage{
calenderweek_from_dates(start = "2017-04-24", end = Sys.Date())
}
\arguments{
\item{start}{start of period (default: '2017-04-24')}

\item{end}{end of period (default: .Date())}
}
\value{
data.frame with daily date sequence for  and corresponding calendar week
}
\description{
Helper function: get calender weeks for time period
}
