% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timezone.R
\name{change_timezone}
\alias{change_timezone}
\title{Timezone change: changes time zone to user defined time zone}
\usage{
change_timezone(df, tz = "UTC", col_datetime = "DateTime",
  debug = TRUE)
}
\arguments{
\item{df}{a dataframe containing a datetime column}

\item{tz}{timezone (default: "UTC")}

\item{col_datetime}{name of the datetime column (default: "DateTime")}

\item{debug}{print debug messages (default: TRUE)}
}
\value{
returns data frame with changed time zone
}
\description{
Timezone change: changes time zone to user defined time zone
}
\references{
Check possible "tz" arguments in column "TZ*" of table
\url{https://en.wikipedia.org/wiki/List_of_tz_database_time_zones} for more
details.
}
