% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_ultimate.R
\name{check_env_nextcloud}
\alias{check_env_nextcloud}
\title{Helper Function: check if all environment variables for Nextcloud are defined}
\usage{
check_env_nextcloud()
}
\value{
TRUE if all defined, FALSE otherwise
}
\description{
Helper Function: check if all environment variables for Nextcloud are defined
}
