% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_thresholds.R
\name{check_thresholds}
\alias{check_thresholds}
\title{Check thresholds}
\usage{
check_thresholds(df, thresholds = get_thresholds())
}
\arguments{
\item{df}{a dataframe as retrieved by import_data_haridwar()}

\item{thresholds}{thresholds dataframe as retrieved by get_thresholds()
(default: "raw")}
}
\value{
dataframe with thresholds check results for selected time period (i.e.
whether Parameters are below/above min/max thresholds defined in dataframe
'thresholds')
}
\description{
Check thresholds
}
