% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_monthly_selection.R
\name{create_monthly_selection}
\alias{create_monthly_selection}
\title{Create monthly selection}
\usage{
create_monthly_selection(startDate = "2016-09-01", endDate = Sys.Date())
}
\arguments{
\item{startDate}{(default: '2016-09-01')}

\item{endDate}{(default: Sys.Date())
(default: "raw")}
}
\value{
dataframe with first/last day for each month between 'startDate' and
'endDate' month including a column 'label' (used in shiny app for month
selection)
}
\description{
Create monthly selection
}
