% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_batch.R
\name{create_report_batch}
\alias{create_report_batch}
\title{Report batch: creates batch file for report}
\usage{
create_report_batch(
  batchDir = file.path(tempdir(), "batch_report"),
  batchName = "create_report.bat",
  report_path = NULL,
  report_config_path = NULL,
  open_in_explorer = TRUE
)
}
\arguments{
\item{batchDir}{path to report batch directory (default: tempdir())}

\item{batchName}{name for report batch file(default: "create_report.bat")}

\item{report_path}{(default: NULL)}

\item{report_config_path}{(default: NULL)}

\item{open_in_explorer}{open batchDir in Windows explorer (default: TRUE).
Only working on a Windows system!}
}
\description{
Report batch: creates batch file for report
}
