% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_operation_berlin_s.R
\name{create_wedeco_metafile}
\alias{create_wedeco_metafile}
\title{Create WEDECO metafile data}
\usage{
create_wedeco_metafile(raw_data_file)
}
\arguments{
\item{raw_data_file}{file path to raw data which should be used for as template
for meta file creation}
}
\value{
data.frame with meta data file structure
}
\description{
Create WEDECO metafile data
}
