% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraph_add_limits.R
\name{dygraph_add_limits}
\alias{dygraph_add_limits}
\title{Dygraph: add (multiple) horizontal lines to plot}
\usage{
dygraph_add_limits(dygraph, limits_df, label_loc = "left",
  col_limits = "ParameterThreshold", col_label = "label", ...)
}
\arguments{
\item{dygraph}{a dygraph object where (possibly) multiple horizontal lines
should be added}

\item{limits_df}{dataframe containing the limits information to be added to
the dygraph (e.g. as retrieved by function get_thresholds())}

\item{label_loc}{Location for horizontal dygraph labels (left or right).
(default: "left)}

\item{col_limits}{column in limits_df containing the limits values (default:
"ParameterThreshold")}

\item{col_label}{column in limits_df containing the label values (default:
"label")}

\item{...}{further arguments passed to dygraphs::dyLimit()}
}
\value{
add limits to existing dygraph object
}
\description{
Dygraph: add (multiple) horizontal lines to plot
}
