% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_and_plot.R
\name{export_data}
\alias{export_data}
\title{CSV data export in "wide" format}
\usage{
export_data(df_long, export_dir, dbg = TRUE)
}
\arguments{
\item{df_long}{data frame in long format (as retrieved by \code{kwb.pilot::group_datetime})}

\item{export_dir}{path to export directory}

\item{dbg}{debug messages (default: TRUE)}
}
\value{
tranforms data in "long" into "wide" format and writes into CSV file
}
\description{
CSV data export in "wide" format
}
