% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_analytics_haridwar.R
\name{fill_datetime}
\alias{fill_datetime}
\title{Helper function: fill empty datetime & create measurementID column in case of
samples}
\usage{
fill_datetime(
  df,
  col_rawData_pattern = "raw",
  col_datetime = "DateTime",
  dbg = FALSE
)
}
\arguments{
\item{df}{a dataframe}

\item{col_rawData_pattern}{column name pattern for identifying raw data
(default: "raw")}

\item{col_datetime}{column name pattern for identifying datetime column
(default: "DateTime"),}

\item{dbg}{print debug messages (default: FALSE)}
}
\value{
returns data frame with filled "DateTime" column and new column
"measurementID" in case of samples
}
\description{
Helper function: fill empty datetime & create measurementID column in case of
samples
}
\keyword{internal}
