% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_ultimate.R
\name{get_env_influxdb_ultimate}
\alias{get_env_influxdb_ultimate}
\title{Helper Function: get influxdb config for Ultimate if defined
defined}
\usage{
get_env_influxdb_ultimate()
}
\value{
list with influxdb config
}
\description{
Helper Function: get influxdb config for Ultimate if defined
defined
}
