% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monthly_data_from_calenderweeks.R
\name{get_monthly_data_from_calendarweeks}
\alias{get_monthly_data_from_calendarweeks}
\title{Helper function for Berlin-S: get all calendar week files for monthy}
\usage{
get_monthly_data_from_calendarweeks(year_month)
}
\arguments{
\item{year_month}{month to be imported (e.g. 2017-04')}
}
\value{
character vector with operational filenames with all calendar weeks
that need to be imported for Berlin Schoenerlinde
}
\description{
Helper function for Berlin-S: get all calendar week files for monthy
}
