% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monthly_periods.R
\name{get_monthly_periods}
\alias{get_monthly_periods}
\title{Get monthly periods}
\usage{
get_monthly_periods(
  year_month_start = "2017-06",
  year_month_end = format(Sys.Date(), "\%Y-\%m"),
  tz = "CET"
)
}
\arguments{
\item{year_month_start}{start year month (default: '2017-06')}

\item{year_month_end}{end year month (default: current month)}

\item{tz}{(default: 'CET')}
}
\value{
dataframe with monthly periods
}
\description{
Get monthly periods
}
