% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_ultimate.R
\name{get_pivot_data}
\alias{get_pivot_data}
\title{InfluxDB: Get Pivot Data from ultimate_mean_ bucket}
\usage{
get_pivot_data(
  agg_interval = "1d",
  date_start = "2021-07-05",
  date_stop = Sys.Date()
)
}
\arguments{
\item{agg_interval}{aggregation interval (default: 1h)}

\item{date_start}{default: "2021-07-05"}

\item{date_stop}{default: Sys.Date()}
}
\value{
pivot data for desired period and aggregation
}
\description{
InfluxDB: Get Pivot Data from ultimate_mean_ bucket
}
