% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_thresholds.R
\name{get_thresholds}
\alias{get_thresholds}
\title{Get thresholds for analytics/operational parameters}
\usage{

  get_thresholds(file = package_file("shiny/haridwar/data/thresholds.csv"))
}
\arguments{
\item{file}{path to csv file with thresholds for Haridwar site (default:
\code{kwb.pilot:::package_file("shiny/haridwar/data/thresholds.csv")})}
}
\value{
returns data frame thresholds for operational/analytical parameters
}
\description{
Get thresholds for analytics/operational parameters
}
