% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timezone.R
\name{get_valid_timezones}
\alias{get_valid_timezones}
\title{Timezone: get valid time zones from Wikipedia}
\usage{
get_valid_timezones()
}
\value{
returns data frame valid time zones (column: TZ.) from Wikipedia
}
\description{
Timezone: get valid time zones from Wikipedia
}
\references{
Check possible "tz" arguments in column "TZ*" of table
\url{https://en.wikipedia.org/wiki/List_of_tz_database_time_zones} for more
details.
}
