% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_datetime.R
\name{group_datetime}
\alias{group_datetime}
\title{Group DateTime by user defined period (year, month, day, hour, minute)}
\usage{
group_datetime(df, by = 600, fun = "stats::median",
  col_datetime = "DateTime", col_datatype = "DataType", dbg = TRUE)
}
\arguments{
\item{df}{a data frame as retrieved by import_data_haridwar()}

\item{by}{an aggregation time step in seconds (default: 600 seconds) for intra-
day aggregation or "day", "month" or "year" for longer time spans}

\item{fun}{function to be used for grouping measurement data of column ParameterValue
(default: stats::median)
(default: system.file("shiny/haridwar/.my.cnf", package = "kwb.pilot"))}

\item{col_datetime}{column name of datetime column (default: DateTime)}

\item{col_datatype}{column name of data type column (default: DataType)}

\item{dbg}{print debug information}
}
\value{
returns data frame with data aggregated according to user defined
aggregation time step
}
\description{
Group DateTime by user defined period (year, month, day, hour, minute)
}
