% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_basel.R
\name{import_analytics_basel}
\alias{import_analytics_basel}
\title{Imports analytical data for Basel (without metadata)}
\usage{
import_analytics_basel(csv_dir = package_file("shiny/basel/data/analytics"))
}
\arguments{
\item{csv_dir}{Define directory with raw analytical data in CSV (.csv) format to
be imported (default: sema.pilot:::package_file("shiny/basel/data/analytics"))}
}
\value{
returns data frame with imported raw analytics data
}
\description{
Imports analytical data for Basel (without metadata)
}
