% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_basel.R
\name{import_analytics_meta_basel}
\alias{import_analytics_meta_basel}
\title{Imports analytical data for Basel (with metadata for both sites at once, i.e.
"rhein" and "wiese")}
\usage{

  import_analytics_meta_basel(analytics_dir = system.file("shiny/basel/data/analytics",
  package = "kwb.pilot"),
  meta_site_path = system.file("shiny/basel/data/metadata/meta_site.csv",
  package = "kwb.pilot"),
  meta_parameter_path = system.file("shiny/basel/data/metadata/meta_parameter.csv",
  package = "kwb.pilot"))
}
\arguments{
\item{analytics_dir}{Define directory with raw analytical data in CSV (.csv) format to
be imported (default: system.file("shiny/basel/data/analytics",
package = "kwb.pilot"))}

\item{meta_site_path}{Define path of "meta_site.csv" to be imported
(default: system.file("shiny/basel/data/metadata/meta_site.csv",
package = "kwb.pilot"))}

\item{meta_parameter_path}{Define path of "meta_parameter.csv" to be imported
(default: system.file("shiny/basel/data/metadata/meta_parameter.csv",
package = "kwb.pilot"))}
}
\value{
data.frame with analytics data for Basel sites including metadata
}
\description{
Imports analytical data for Basel (with metadata for both sites at once, i.e.
"rhein" and "wiese")
}
