% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_berlin_f.R
\name{import_data_berlin_f}
\alias{import_data_berlin_f}
\title{Import data for Berlin Friedrichshagen}
\usage{
import_data_berlin_f(
  raw_data_files = fs::dir_ls(package_file("shiny/berlin_f/data/raw/online_data"),
    recurse = TRUE, regexp = "^[^~].*\\\\.xlsx$"),
  meta_file_path =
    package_file("shiny/berlin_f/data/raw/online_data/parameter_unit_metadata.csv")
)
}
\arguments{
\item{raw_data_files}{vector with full path to operational raw data files that
allows to limit import to specific files (default:
fs::dir_ls(package_file("shiny/berlin_f/data/raw/online_data""), recurse = TRUE,
regexp = "^[^~].*\\.xlsx$")).}

\item{meta_file_path}{path to metadata file (default:
kwb.pilot:::package_file(file.path("shiny/berlin_f/data/raw/online_data",
"parameter_site_metadata.csv"))}
}
\value{
data.frame with imported operational data (analytics´data to be added as
soon as available)
}
\description{
Import data for Berlin Friedrichshagen
}
