% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_operation_berlin_t.R
\name{import_lab_data_berlin_t}
\alias{import_lab_data_berlin_t}
\title{BerlinTiefwerder: import lab data}
\usage{
import_lab_data_berlin_t(
  xlsx_path = shiny_file("berlin_t/data/analytics.xlsx")
)
}
\arguments{
\item{xlsx_path}{full path to lab data EXCEL file in xlsx format
(default: kwb.pilot:::shiny_file("berlin_t/data/analytics.xlsx"))}
}
\value{
a list of imported lab data for Berlin-Tiefwerder
}
\description{
BerlinTiefwerder: import lab data
}
