% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_operation_haridwar.R
\name{import_operation}
\alias{import_operation}
\title{Imports operational data}
\usage{
import_operation(mysql_conf = file.path(getwd(), ".my.cnf"))
}
\arguments{
\item{mysql_conf}{path to the MySQL configuration file}
}
\value{
returns data frame operational data from MySQL db
}
\description{
Imports operational data
}
