% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_basel.R
\name{import_operation_basel}
\alias{import_operation_basel}
\title{Imports operational data for Basel (without metadata and only for one site
at once, e.g. "rhein" or "wiese")}
\usage{

  import_operation_basel(xlsx_dir = system.file("shiny/basel/data/operation/wiese",
  package = "kwb.pilot"))
}
\arguments{
\item{xlsx_dir}{Define directory with raw data in EXCEL spreadsheet (.xlsx) to
be imported (default: system.file("shiny/basel/data/operation/wiese",
package = "kwb.pilot"))}
}
\value{
returns data frame with imported raw operational data
}
\description{
Imports operational data for Basel (without metadata and only for one site
at once, e.g. "rhein" or "wiese")
}
