% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_basel.R
\name{import_operation_meta_basel}
\alias{import_operation_meta_basel}
\title{Imports operational data for Basel (with metadata for
both sites at once, i.e. "rhein" and "wiese")}
\usage{

  import_operation_meta_basel(raw_dir_rhein = system.file(file.path("shiny",
  "basel/data/operation/rhein"), package = "kwb.pilot"),
  raw_dir_wiese = system.file("shiny/basel/data/operation/wiese", package
  = "kwb.pilot"),
  meta_online_path = system.file("shiny/basel/data/metadata/meta_online.csv",
  package = "kwb.pilot"),
  meta_site_path = system.file("shiny/basel/data/metadata/meta_site.csv",
  package = "kwb.pilot"),
  meta_parameter_path = system.file("shiny/basel/data/metadata/meta_parameter.csv",
  package = "kwb.pilot"))
}
\arguments{
\item{raw_dir_rhein}{Define directory for site "rhein" with raw data in
EXCEL spreadsheet format (.xlsx) to be imported (default:
system.file("shiny/basel/data/operation/rhein", package = "kwb.pilot"))}

\item{raw_dir_wiese}{Define directory for site "rhein" with raw data in
EXCEL spreadsheet format (.xlsx) to be imported (default:
system.file("shiny/basel/data/operation/wiese", package = "kwb.pilot"))}

\item{meta_online_path}{path to file containing metadata for online data
(default: system.file("shiny/basel/data/metadata/meta_online.csv",
package = "kwb.pilot"))}

\item{meta_site_path}{Define path of "meta_site.csv" to be imported
(default: system.file("shiny/basel/data/metadata/meta_site.csv",
package = "kwb.pilot"))}

\item{meta_parameter_path}{Define path of "meta_parameter.csv" to be imported
(default: system.file("shiny/basel/data/metadata/meta_parameter.csv",
package = "kwb.pilot"))}
}
\value{
returns data frame with imported raw operational data with metadata
for both sites (i.e."rhein" and "wiese")

data.frame with operational data for Basel sites including metadata
}
\description{
Imports operational data for Basel (with metadata for
both sites at once, i.e. "rhein" and "wiese")
}
