% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_analytics_haridwar.R
\name{import_sheet}
\alias{import_sheet}
\title{Imports an analytics sheet from an EXCEL spreadsheet}
\usage{
import_sheet(xlsPath, sheet, col_names = TRUE,
  col_rawData_pattern = "raw",
  col_ignore_pattern = "mean|empty|X_|RX|not_used", skip = 69,
  tz_org = "UTC", tz_export = "UTC")
}
\arguments{
\item{xlsPath}{path to xls file with analytics data}

\item{sheet}{a sheet in spreadsheet defined with "xlsPath" containing
analytics data (check with: readxl::excel_sheets(xlsPath))}

\item{col_names}{do the (default: TRUE)
(default: "DateTime"),}

\item{col_rawData_pattern}{specify pattenr of columns containing raw data
(default: "raw")}

\item{col_ignore_pattern}{specify pattern of columns that should be ignored
of importing (default: "mean|empty|X_|RX|not_used")}

\item{skip}{number of rows in sheet to skip (default: 69),}

\item{tz_org}{specify timezone of samples  (default: "UTC")}

\item{tz_export}{specify timezone for data export (default: "UTC")}
}
\value{
returns data frame with normalised analytics data in list form
}
\description{
Imports an analytics sheet from an EXCEL spreadsheet
}
\keyword{internal}
