% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_analytics_haridwar.R
\name{import_sheets}
\alias{import_sheets}
\title{Imports multiple analytics sheets from an EXCEL spreadsheet}
\usage{
import_sheets(xlsPath, sheets_analytics, sheet_parameters = "Parameters",
  sheet_sites = "Sites", sheet_location = "Location",
  col_rawData_pattern = "raw",
  col_ignore_pattern = "mean|empty|X_|RX|not_used", skip = 69,
  tz_org = NULL, tz_export = "UTC", dbg = TRUE)
}
\arguments{
\item{xlsPath}{path to xls file with analytics data}

\item{sheets_analytics}{a character vector with the names of the sheets
with analytics data (check with: readxl::excel_sheets(xlsPath))}

\item{sheet_parameters}{sheet name containing parameter metadata (default:
"Parameters")}

\item{sheet_sites}{sheet name containing sites metadata (default:
"Sites")}

\item{sheet_location}{sheet name containing location metadata (default:
"Location")}

\item{col_rawData_pattern}{specify pattenr of columns containing raw data
(default: "raw")}

\item{col_ignore_pattern}{specify pattern of columns that should be ignored
of importing (default: "mean|empty|X_|RX|not_used")}

\item{skip}{number of rows in sheet to skip (default: 69),}

\item{tz_org}{specify timezone of samples  (default: "UTC")}

\item{tz_export}{specify timezone for data export (default: "UTC")}

\item{dbg}{print debug messages (default: TRUE)}
}
\value{
returns data frame with normalised analytics data in list form
}
\description{
Imports multiple analytics sheets from an EXCEL spreadsheet
}
