% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timezone.R
\name{is_POSIXct}
\alias{is_POSIXct}
\title{Helper function: to check whether object is of type POSIXct}
\usage{
is_POSIXct(x)
}
\arguments{
\item{x}{a vector that should be tested whether}
}
\value{
returns TRUE if of tpye POSIXct
}
\description{
Helper function: to check whether object is of type POSIXct
}
\keyword{internal}
