% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monthly_periods.R
\name{last_day}
\alias{last_day}
\title{Helper function to get last day of month}
\usage{
last_day(date)
}
\arguments{
\item{date}{in format as.Date()}
}
\value{
vector with last day of month(s)
}
\description{
Helper function to get last day of month
}
\keyword{internal}
