% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fst.R
\name{load_fst_data}
\alias{load_fst_data}
\title{Load fst data for shiny app}
\usage{
load_fst_data(fst_dir)
}
\arguments{
\item{fst_dir}{directory of fst files to be loaded}
}
\description{
Load fst data for shiny app
}
