% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_and_plot.R
\name{long_to_wide}
\alias{long_to_wide}
\title{Helper function: transform "long" to "wide"}
\usage{
long_to_wide(df)
}
\arguments{
\item{df}{data frame in long format (as retrieved by \code{kwb.pilot::group_datetime})}
}
\description{
Helper function: transform "long" to "wide"
}
