% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mbr4.0_metadata}
\alias{mbr4.0_metadata}
\title{Metadata: MBR 4.0 pilot plant Berlin}
\format{
A data frame with 78 rows and 10 variables
}
\usage{
mbr4.0_metadata
}
\description{
A dataset with the following columns:  ParameterCode_SiteCode, 
ParameterCode, SiteCode, ParameterName_SiteName, ParameterName, 
SiteName, ParameterUnit, Comment, Source, DataType
}
\keyword{datasets}
