% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_and_export_fst.R
\name{merge_and_export_fst}
\alias{merge_and_export_fst}
\title{Helper function: merge and export fst files into main shiny data folder}
\usage{
merge_and_export_fst(time_pattern = NULL, compression = 100,
  import_dir = system.file("shiny/berlin_t/data/fst", package =
  "kwb.pilot"), export_dir = system.file("shiny/berlin_t/data", package =
  "kwb.pilot"))
}
\arguments{
\item{time_pattern}{optional pattern to filter months to be imported (default: NULL),
for using it do e.g. "2017-06|2017-07" or c("2017-06", "2017-07")}

\item{compression}{compression for fst export (default: 100)}

\item{import_dir}{directory with fst files or subdirs to be imported (default:
system.file("shiny/berlin_t/data/fst",package = "kwb.pilot"))}

\item{export_dir}{directory with fst directory for export (default:
system.file("shiny/berlin_t/data",package = "kwb.pilot"))}
}
\value{
imports multiple fst files and exports them to be used for app
}
\description{
Helper function: merge and export fst files into main shiny data folder
}
