% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_nextcloud_files.R
\name{move_nextcloud_files}
\alias{move_nextcloud_files}
\title{Move Nextcloud Files}
\usage{
move_nextcloud_files(paths_df, overwrite = FALSE, dbg = TRUE)
}
\arguments{
\item{paths_df}{dataframe with columns \code{raw} and \code{imported} with
full path to source and target file}

\item{overwrite}{overwrite files if these exist in target directory (default:
TRUE)}

\item{dbg}{debug messages (default: TRUE)}
}
\value{
moves file to target file
}
\description{
Move Nextcloud Files
}
