% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_operational_parameters_berlin_f.R
\name{normalised_permeate_flow}
\alias{normalised_permeate_flow}
\title{Calculate normalised permeate flow}
\usage{
normalised_permeate_flow(
  tempFeed,
  conLoop,
  vfrPerm,
  vfrLoop,
  vfrFeed,
  prePerm,
  preProc,
  preConc,
  nwp0 = 1.429162,
  vfrPerm0 = 800
)
}
\arguments{
\item{tempFeed}{tempFeed}

\item{conLoop}{conLoop}

\item{vfrPerm}{vfrPerm}

\item{vfrLoop}{vfrLoop}

\item{vfrFeed}{vfrFeed}

\item{prePerm}{prePerm}

\item{preProc}{preProc}

\item{preConc}{preConc}

\item{nwp0}{nwp0}

\item{vfrPerm0}{vfrPerm0}
}
\value{
nwpt
}
\description{
Calculate normalised permeate flow
}
