% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_analytics_haridwar.R
\name{plot_analytics}
\alias{plot_analytics}
\title{Plot analytics data (in PDF)}
\usage{
plot_analytics(df)
}
\arguments{
\item{df}{dataframe as retrieved by import_sheets()}
}
\value{
creates new subdirectory "/report" in current working directory
and stores pdf plots there
}
\description{
Plot analytics data (in PDF)
}
