% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_operational_parameters_haridwar.R
\name{plot_calculated_operational_timeseries}
\alias{plot_calculated_operational_timeseries}
\title{Plot calculate operational time series}
\usage{
plot_calculated_operational_timeseries(df)
}
\arguments{
\item{df}{a data frame as retrieved by calculate_operational_parameters()}
}
\value{
plots time series for calculated operational parameters
}
\description{
Plot calculate operational time series
}
\examples{
\dontrun{
haridwar_raw_list <- import_data_haridwar()
myDat <- calculate_operational_parameters(df = haridwar_raw_list)
plot_calculated_operational_timeseries(myDat)}
}
