% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_and_plot.R
\name{plot_data}
\alias{plot_data}
\title{Export interactive HTML plot with "plotly"}
\usage{
plot_data(df_long, export_dir, dbg = TRUE)
}
\arguments{
\item{df_long}{data frame in long format (as retrieved by \code{kwb.pilot::group_datetime})}

\item{export_dir}{path to export directory}

\item{dbg}{debug messages (default: TRUE)}
}
\value{
interactive HTML plots in subdirectory "<export_dir>/plots/"
}
\description{
Export interactive HTML plot with "plotly"
}
