% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fst.R
\name{read_fst}
\alias{read_fst}
\title{Wrapper for fst::read.fst to read DateTime column in POSIXct format}
\usage{
read_fst(path, tz = "CET", col_datetime = "DateTime", ...)
}
\arguments{
\item{path}{path to fst file}

\item{tz}{timezone of DateTime to be imported (default: "CET")}

\item{col_datetime}{column name containing numeric values in nanoseconds since
1970-01-01 (default: "DateTime")}

\item{...}{further arguments passed to fst::read.fst}
}
\value{
data.frame with formatting of DateTime column POSIXct
}
\description{
Wrapper for fst::read.fst to read DateTime column in POSIXct format
}
