% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_martin_systems.R
\name{read_martin_systems}
\alias{read_martin_systems}
\title{Read Martin Systems data from single CSV file}
\usage{
read_martin_systems(path, tz = "CET", dbg = TRUE)
}
\arguments{
\item{path}{path to Martin Systems file}

\item{tz}{time zone (default: CET) the measurements are taken (passed to 
function readr::locale(tz = tz))}

\item{dbg}{debug (default: TRUE)}
}
\value{
data frame with Martin Systems raw data in "long" format
}
\description{
CSV format should use the ";" as field and "," as decimal 
separator
}
