% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mbr4.R
\name{read_mbr4}
\alias{read_mbr4}
\title{Read MBR4.0 data combining latest and archived data}
\usage{
read_mbr4(
  latest_url = Sys.getenv("MBR40_URL"),
  archived_file = "MBR_export_",
  archived_dir = "projects/MBR4.0/Exchange/Rohdaten/Online_export",
  archived_url = Sys.getenv("NEXTCLOUD_URL"),
  archived_user = Sys.getenv("NEXTCLOUD_USER"),
  archived_pw = Sys.getenv("NEXTCLOUD_USER"),
  target_dir = tempdir(),
  dbg = FALSE,
  ...
)
}
\arguments{
\item{latest_url}{url to download latest .tsv file (default: Sys.getenv("MBR40_URL"),
please use run \link[usethis]{edit_r_environ} and define a
key value pair "MBR40_URL" = "download-url-martin-systems") so
that this function works automatically}

\item{archived_file}{file name/pattern of XLSX file  (default: "MBR_export_")}

\item{archived_dir}{directory on Nextcloud containing file (default: 
"projects/MBR4.0/Exchange/Rohdaten/Online_export")}

\item{archived_url}{url of Nextcloud (default: Sys.getenv("NEXTCLOUD_URL"))}

\item{archived_user}{username of Nextcloud  (default: Sys.getenv("NEXTCLOUD_USER"))}

\item{archived_pw}{password of Nextcloud  (default: Sys.getenv("NEXTCLOUD_USER"))}

\item{target_dir}{directory to download data (default: tempdir())}

\item{dbg}{print debug messages (default: FALSE)}

\item{...}{additional arguments passed to  \link[readr]{read_tsv}}
}
\value{
tibble with imported MBR4.0 xlsx data (archived on Nextcloud since 
start of operation)
}
\description{
Download latest data as 'tsv' from Martin Systems Webportal and 
combine with archived ('tsv') on Nextcloud'
}
\examples{
\dontrun{
mbr4_data <- read_mbr4()
str(mbr4_data)
}
}
