% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mbr4.R
\name{read_mbr4_archived}
\alias{read_mbr4_archived}
\title{Read MBR4.0 archived data from Nextcloud}
\usage{
read_mbr4_archived(
  file = "MBR_export_",
  dir = "projects/MBR4.0/Exchange/Rohdaten/Online_export",
  target_dir = tempdir(),
  url = Sys.getenv("NEXTCLOUD_URL"),
  user = Sys.getenv("NEXTCLOUD_USER"),
  pw = Sys.getenv("NEXTCLOUD_USER"),
  dbg = FALSE,
  ...
)
}
\arguments{
\item{file}{file name/pattern of tsv file  (default: "MBR_export_")}

\item{dir}{directory on Nextcloud containing file (default: 
"projects/MBR4.0/Exchange/Rohdaten/Online_export")}

\item{target_dir}{directory to download data (default: tempdir())}

\item{url}{url of Nextcloud (default: Sys.getenv("NEXTCLOUD_URL"))}

\item{user}{username of Nextcloud  (default: Sys.getenv("NEXTCLOUD_USER"))}

\item{pw}{password of Nextcloud  (default: Sys.getenv("NEXTCLOUD_USER"))}

\item{dbg}{print debug messages (default: FALSE)}

\item{...}{additional arguments passed to  \link[readr]{read_tsv}}
}
\value{
tibble with imported archived MBR4.0 xlsx data from Nextcloud
}
\description{
Read MBR4.0 archived data from Nextcloud
}
\examples{
mbr4_data_archived <- read_mbr4_archived()
str(mbr4_data_archived)
}
