% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mbr4.R
\name{read_mbr4_latest}
\alias{read_mbr4_latest}
\title{Read MBR4.0 data from Martin Systems Webportal (As "tsv")}
\usage{
read_mbr4_latest(
  url = Sys.getenv("MBR40_URL"),
  target_dir = tempdir(),
  locale = readr::locale(tz = "CET", decimal_mark = ".", grouping_mark = ","),
  dbg = FALSE,
  ...
)
}
\arguments{
\item{url}{url to download site (default: Sys.getenv("MBR4.0_URL"),
please use run \link[usethis]{edit_r_environ} and define a
key value pair "MBR40_URL" = "download-url-martin-systems") so
that this function works automatically}

\item{target_dir}{directory to download data (default: tempdir())}

\item{locale}{locale (default: \code{\link[readr]{locale}}(tz = "CET",
decimal_mark = ".", grouping_mark = ","))}

\item{dbg}{print debug messages (default: FALSE)}

\item{...}{additional arguments passed to  \link[readr]{read_tsv}}
}
\value{
tibble with imported MBR4.0 tsv data (~ last four weeks)
}
\description{
Read MBR4.0 data from Martin Systems Webportal (As "tsv")
}
\examples{
mbr4_data_latest <- read_mbr4_latest()
str(mbr4_data_latest)
}
