% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mbr4.R
\name{read_mbr4_tsv}
\alias{read_mbr4_tsv}
\title{Read MBR4.0 tsv data}
\usage{
read_mbr4_tsv(path, dbg = FALSE, ...)
}
\arguments{
\item{path}{path to tsv file to be imported}

\item{dbg}{print debug messages (default: FALSE)}

\item{...}{additional arguments passed to  \link[readr]{read_tsv}}
}
\value{
Reads MBR4.0 tsv data
}
\description{
Read MBR4.0 tsv data
}
