% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_operation_berlin_t.R
\name{read_pentair_data}
\alias{read_pentair_data}
\title{Read PENTAIR operational data}
\usage{

  read_pentair_data(raw_data_dir = system.file("shiny/berlin_t/data/operation",
  package = "kwb.pilot"), raw_data_files = NULL,
  meta_file_path = system.file("shiny/berlin_t/data/parameter_site_metadata.csv",
  package = "kwb.pilot"))
}
\arguments{
\item{raw_data_dir}{path of directory containing PENTAIR xls files (default:
(default: system.file("shiny/berlin_t/data/operation",
package = "kwb.pilot"))))}

\item{raw_data_files}{vector with full path to operational raw data files that
allows to limit import to specific files (default: NULL). If specified parameter
"raw_data_dir" will not be used}

\item{meta_file_path}{path to metadata file (default:
system.file("shiny/berlin_t/data/parameter_site_metadata.csv", package =
"kwb.pilot")))}
}
\value{
data.frame with imported PENTAIR operational data
}
\description{
Read PENTAIR operational data
}
