% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_weintek.R
\name{read_weintek}
\alias{read_weintek}
\title{Read Weintek data from single file}
\usage{
read_weintek(path, tz = "CET", dbg = TRUE)
}
\arguments{
\item{path}{path to Weintek file}

\item{tz}{time zone (default: CET) the measurements are taken (passed to 
function kwb.pilot:::set_timezone())}

\item{dbg}{debug (default: TRUE)}
}
\value{
data frame with Weintek raw data
}
\description{
Read Weintek data from single file
}
