% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_weintek.R
\name{read_weintek_batch}
\alias{read_weintek_batch}
\title{Read Weintek data from multiple files}
\usage{
read_weintek_batch(files, tz = "CET", dbg = TRUE)
}
\arguments{
\item{files}{path to Weintek files}

\item{tz}{time zone (default: CET) the measurements are taken (passed to
function kwb.pilot:::set_timezone())}

\item{dbg}{debug (default: TRUE)}
}
\value{
data frame with Weintek raw data
}
\description{
Read Weintek data from multiple files
}
