% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_configuration.R
\name{report_config_template}
\alias{report_config_template}
\title{Report config: generate template}
\usage{
report_config_template(df = NULL, temporal_aggregation = "raw",
  output_timezone = "UTC")
}
\arguments{
\item{df}{a dataframe as retrieved by import_data_haridwar()}

\item{temporal_aggregation}{Set the following values if data should be
summarised to e.g. 10 minutes (600) or hourly (3600), daily ("day")
or monthly ("month") median values
(default: "raw")}

\item{output_timezone}{into which timezone should the data be converted for
the report? (default: "UTC")}
}
\value{
default list for report configuration template
}
\description{
Report config: generate template
}
