% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_configuration.R
\name{report_txt_to_config}
\alias{report_txt_to_config}
\title{Report config: imports text file to list}
\usage{
report_txt_to_config(config_txt = "report_config.txt")
}
\arguments{
\item{config_txt}{path to report configuration text file created by
a report configuration list e.g. as retrieved by function report_config_to_txt()}
}
\value{
saves report configuration list as text file
}
\description{
Report config: imports text file to list
}
\examples{
\dontrun{
### Creates a configuration template
config <- report_config_template()
### Saves list config in text
report_config_to_txt(config_list = config, output_file = "report_config.txt")
### Reads config list from text file to
config_imported <- report_txt_to_config(config_txt = "report_config.txt")
### Check whether both are identical
identical(x = config, y = config_imported)
}
}
