% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{run_app}
\alias{run_app}
\title{Runs Shiny app for an AQUANES site}
\usage{
run_app(siteName = "haridwar", use_live_data = FALSE,
  mySQL_conf = NULL, launch.browser = TRUE, ...)
}
\arguments{
\item{siteName}{site name for shiny app (default: "haridwar")}

\item{use_live_data}{should live data be used (default: FALSE)}

\item{mySQL_conf}{file path to mySQL config file (.my.cnf). Only used if
parameter use_live_data is TRUE and there is no .my.cnf in the app folder for
the selected site (default: NULL)}

\item{launch.browser}{If true, the system's default web browser will be
launched automatically after the app is started (default: TRUE)}

\item{...}{further arguments passed to shiny::runApp()}
}
\description{
Runs Shiny app for an AQUANES site
}
