% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timezone.R
\name{set_timezone}
\alias{set_timezone}
\title{Timezone set: sets a user defined time zone}
\usage{
set_timezone(df, tz = "UTC", col_datetime = "DateTime")
}
\arguments{
\item{df}{a dataframe containing a datetime column}

\item{tz}{timezone (default: "UTC")}

\item{col_datetime}{name of the datetime column (default: "DateTime")}
}
\value{
returns data frame with specified time zone
}
\description{
Timezone set: sets a user defined time zone
}
\references{
Check possible "tz" arguments in column "TZ*" of table
\url{https://en.wikipedia.org/wiki/List_of_tz_database_time_zones} for more
details.
}
