% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shiny_file}
\alias{shiny_file}
\title{Path to Shiny File in Package}
\usage{
shiny_file(...)
}
\arguments{
\item{...}{relative path to file in "shiny" folder of R package "kwb.pilot"}
}
\value{
path to file (if existing)
}
\description{
Path to Shiny File in Package
}
\examples{
shiny_file()
}
