% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_operation_haridwar.R
\name{src_mysql_from_cnf}
\alias{src_mysql_from_cnf}
\title{Helper function: store MySQL database credentials in ".my.cnf"}
\usage{
src_mysql_from_cnf(dbname, group = NULL, dir = file.path(getwd(),
  ".my.cnf"), host = NULL, user = NULL, password = NULL, ...)
}
\arguments{
\item{dbname}{name of database to be accessed}

\item{group}{name of group}

\item{dir}{path to the MySQL configuration file (default: file.path(getwd(),
".my.cnf"))}

\item{host}{hostname of MySQL database (default: NULL)}

\item{user}{username of MySQL database (default: NULL)}

\item{password}{password of MySQL database (default: NULL)}

\item{...}{further arguments passed to dplyr::src_mysql()}
}
\value{
sets dplyr::src_mysql to work with MySQL config file
}
\description{
Helper function: store MySQL database credentials in ".my.cnf"
}
\keyword{internal}
