% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_mbr4_data.R
\name{tidy_mbr4_data}
\alias{tidy_mbr4_data}
\title{MBR 4.0 Data Tidy}
\usage{
tidy_mbr4_data(
  mbr4_data,
  path_metadata = kwb.pilot:::shiny_file("mbr4.0/data/metadata.csv")
)
}
\arguments{
\item{mbr4_data}{tibble as retrieved by \link{read_mbr4}}
}
\value{
tidy MBR 4 data in long format
}
\description{
MBR 4.0 Data Tidy
}
\examples{
\dontrun{
mbr4_data <- read_mbr4()
mbr4_data_tidy <- kwb.pilot::tidy_mbr4_data(mbr4_data)
}
}
