% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_ultimate.R
\name{write_aggr_to_influxdb}
\alias{write_aggr_to_influxdb}
\title{InfluxDB: write aggregated time series to Ultimate target bucket}
\usage{
write_aggr_to_influxdb(
  start,
  end,
  agg_interval = "1h",
  agg_function = "mean",
  bucket_source = "ultimate",
  bucket_target = sprintf("\%s_\%s_\%s", bucket_source, agg_function, agg_interval),
  bucket_org = "kwb"
)
}
\arguments{
\item{start}{date start}

\item{end}{date end}

\item{agg_interval}{aggregation interval (default: "1h")}

\item{agg_function}{aggregation function (default: "mean")}

\item{bucket_source}{bucket source (default: "ultimate")}

\item{bucket_target}{bucket target (default: <bucket_source_<agg_function>_<agg_interval>))}

\item{bucket_org}{bucket organisation (default: "kwb")}

\item{max_days}{maximum time period in days that should be sent within one
query to influxdb (default: 5)}
}
\value{
writes aggregated time series to InfluxDB target bucket in loop
}
\description{
InfluxDB: write aggregated time series to Ultimate target bucket
}
