% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_ultimate.R
\name{write_aggr_to_influxdb_loop}
\alias{write_aggr_to_influxdb_loop}
\title{InfluxDB: write aggregated time series to Ultimate target bucket in loop}
\usage{
write_aggr_to_influxdb_loop(
  agg_period = "1h",
  agg_function = "median",
  bucket_source = "ultimate",
  bucket_target = sprintf("\%s_\%s_\%s", bucket_source, agg_function, agg_period),
  bucket_org = "kwb",
  max_days = 5
)
}
\arguments{
\item{agg_function}{aggregation function (default: "mean")}

\item{bucket_source}{bucket source (default: "ultimate")}

\item{bucket_target}{bucket target (default: <bucket_source_<agg_function>_<agg_interval>))}

\item{bucket_org}{bucket organisation (default: "kwb")}

\item{max_days}{maximum time period in days that should be sent within one
query to influxdb (default: 5)}

\item{agg_interval}{aggregation interval (default: "1h")}

\item{date_start}{date start (default: "2021-07-05")}

\item{date_end}{date end (default: Sys.Date())}

\item{hour_start}{(default: 0)}

\item{hour_end}{(default: 12)}
}
\value{
writes aggregated time series to InfluxDB target bucket in loop
}
\description{
wrapper for \code{\link{write_aggr_to_influxdb }}
}
