% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_batch.R
\name{write_report}
\alias{write_report}
\title{Report batch: creates batch file for report}
\usage{
write_report(rmd_name = "report.Rmd", output_dir = getwd())
}
\arguments{
\item{rmd_name}{name of Rmarkdown document to render (default: "report.Rmd")}

\item{output_dir}{outputDirectory (default: getwd())}
}
\description{
Report batch: creates batch file for report
}
\keyword{internal}
