% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_ultimate.R
\name{write_to_influxdb}
\alias{write_to_influxdb}
\title{InfluxDB: write to InfluxDB}
\usage{
write_to_influxdb(tsv_paths, paths, batch_size = 5000)
}
\arguments{
\item{tsv_paths}{vector with tsv_paths with files to be imported by
a modification of \code{\link{read_pentair_data}}}

\item{paths}{paths list with elements \code{raw_data_dir} and \code{site_code}}

\item{batch_size}{number of data points that are written in one query (default:
5000)}
}
\value{
writes imported data to InfluxDB
}
\description{
InfluxDB: write to InfluxDB
}
