% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_ultimate.R
\name{write_to_influxdb_loop}
\alias{write_to_influxdb_loop}
\title{InfluxDB: write to InfluxDB in Loop}
\usage{
write_to_influxdb_loop(
  tsv_paths,
  paths,
  changed_only = TRUE,
  max_tsv_files = 5,
  batch_size = 5000
)
}
\arguments{
\item{tsv_paths}{vector with tsv_paths with files to be imported by
\code{\link{write_to_influxdb}} which relies on \code{\link{read_pentair_data}}}

\item{paths}{paths list with elements \code{raw_data_dir} and \code{site_code}}

\item{changed_only}{TRUE if only columns with changing data points 
within time series of provided tsv_paths (limited by parameter \code{max_tsv_files},
i.e. changes between different tsv splits are not detected!) should be written 
to InfluxDB, otherwise FALSE (default: TRUE)}

\item{max_tsv_files}{maximum number of tsv files to read at once (should be
limited due to high RAM demand), default: 5}

\item{batch_size}{number of data points that are written in one query (default:
5000)}
}
\value{
writes imported data to InfluxDB in Loop
}
\description{
wrapper function for \code{\link{write_to_influxdb}}
}
