% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_basel.R
\name{add_site_metadata}
\alias{add_site_metadata}
\title{Helper function: add site metadata}
\usage{
add_site_metadata(
  df,
  df_col_sitecode = "SiteCode",
  meta_site_path = shiny_file("basel/data/metadata/meta_site.csv")
)
}
\arguments{
\item{df}{data frame containing at least a column "SiteCode"}

\item{df_col_sitecode}{column in df containing site code (default: "SiteCode")}

\item{meta_site_path}{Define path of "meta_site.csv" to be imported
(default: sema.pilot:::shiny_file("basel/data/metadata/meta_site.csv"))}
}
\value{
returns input data frame with joined metadata
}
\description{
Helper function: add site metadata
}
