% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_operational_parameters_berlin_f.R
\name{calculate_operational_parameters_berlin_f}
\alias{calculate_operational_parameters_berlin_f}
\title{Calculate operational parameters for Berlin-Friedrichshagen}
\usage{
calculate_operational_parameters_berlin_f(
  df,
  calc_list = get_calc_info_from_config(config, "expr"),
  calc_list_name = get_calc_info_from_config(config, "name"),
  calc_list_unit = get_calc_info_from_config(config, "unit"),
  calc_paras = get_calc_info_from_config(config, "paras"),
  config = get_calc_config("berlin_f")
)
}
\arguments{
\item{df}{a data frame as retrieved by import_data_berlin_f()}

\item{calc_list}{list with calculation operations to be carried out}

\item{calc_list_name}{full names of parameters to be used for plotting for
each calculation specified wit 'calc_list'.}

\item{calc_list_unit}{units of parameters to be used for plotting for each
calculation specified wit 'calc_list'.}

\item{calc_paras}{a vector with parameter codes used for performing
calculations defined in 'calc_list'}

\item{config}{configuration object (list) from which the \code{calc_*}
arguments are filled. Default: \code{get_calc_config("berlin_f")}}
}
\value{
dataframe with calculated operational parameters
}
\description{
Calculate operational parameters for Berlin-Friedrichshagen
}
\examples{
\dontrun{
raw_list <- import_data_berlin_f()
myDat <- calculate_operational_parameters_berlin_f(df = raw_list)
}

}
