% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_operational_parameters_berlin_t.R
\name{calculate_operational_parameters_berlin_t}
\alias{calculate_operational_parameters_berlin_t}
\title{Calculate operational parameters for Berlin-Tiefwerder}
\usage{
calculate_operational_parameters_berlin_t(
  df,
  calc_list = list(recovery = "100*`FY-20-01`/`FT-10-01`"),
  calc_list_name = c("recovery"),
  calc_list_unit = c("\%"),
  calc_paras = c("FY-20-01", "FT-10-01")
)
}
\arguments{
\item{df}{a data frame as retrieved by read_pentair_data()}

\item{calc_list}{list with calculation operations to be carried out
(default: list(recovery  = "100*`FY-20-01`/`FT-10-01`"))}

\item{calc_list_name}{full names of parameters to be used for plotting for each
calculation specified wit 'calc_list'. default: c('recovery')}

\item{calc_list_unit}{units of parameters to be used for plotting for each
calculation specified wit 'calc_list'. default: c("percent")}

\item{calc_paras}{a vector with parameter codes used for performing calculations
defined in 'calc_list' (default: c("FY-20-01", "FT-10-01")}
}
\value{
dataframe with calculated operational parameters
}
\description{
Calculate operational parameters for Berlin-Tiefwerder
}
\examples{
\dontrun{
raw_list <- read_pentair_data()
myDat <- calculate_operational_parameters_berlin_t(df = raw_list)
}

}
