% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_haridwar.R
\name{import_data_haridwar}
\alias{import_data_haridwar}
\title{Imports Haridwar data}
\usage{

  import_data_haridwar(analytics_path = package_file("shiny/haridwar/data/analytics.xlsx"),
  operation_mySQL_conf = package_file("shiny/haridwar/.my.cnf"),
  operation_meta_path = package_file("shiny/haridwar/data/operation_parameters.csv"),
  excludedSheets = c("Parameters", "Location", "Sites", "#Summary",
  "Site_and_Parameter", "Observations", "dP", "ORP", "Flow",
  "Current_Voltage", "As_total_Arsenator"), skip = 69, debug = TRUE)
}
\arguments{
\item{analytics_path}{Define path of analytics EXCEL spreadsheet to be
imported (default: kwb.pilot:::package_file("shiny/haridwar/data/analytics.xlsx"))}

\item{operation_mySQL_conf}{column name pattern for identifying raw data
(default: kwb.pilot:::package_file("shiny/haridwar/.my.cnf"))}

\item{operation_meta_path}{path to table with meta data for operational
parameters (default: kwb.pilot:::package_file("shiny/haridwar/data/operation_parameters.csv"))}

\item{excludedSheets}{all sheets, which are not listed here will be imported
as lab data sheets (default: c("Parameters", "Location", "Sites", "#Summary",
"Site_and_Parameter", "Observations", "dP", "ORP", "Flow", "Current_Voltage",
"As_total_Arsenator"))}

\item{skip}{number of rows to skip for each lab data sheet (default: 69), i.e.
for all sheets which are not explictly excluded with parameter "excludedSheets"}

\item{debug}{if TRUE print debug messages (default: TRUE)}
}
\value{
returns data frame with Haridwar raw data (operation & analytics)
}
\description{
Imports Haridwar data
}
