% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_operation_berlin_t.R
\name{import_lab_data_berlin_t}
\alias{import_lab_data_berlin_t}
\title{BerlinTiefwerder: import lab data}
\usage{

  import_lab_data_berlin_t(xlsx_path = system.file("shiny/berlin_t/data/analytics.xlsx",
  package = "kwb.pilot"))
}
\arguments{
\item{xlsx_path}{full path to lab data EXCEL file in xlsx format (default:
(default: system.file("shiny/berlin_t/data/analytics.xlsx",
package = "kwb.pilot"))))}
}
\value{
a list of imported lab data for Berlin-Tiefwerder
}
\description{
BerlinTiefwerder: import lab data
}
