% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_basel.R
\name{import_operation_meta_basel}
\alias{import_operation_meta_basel}
\title{Imports operational data for Basel (with metadata for
both sites at once, i.e. "rhein" and "wiese")}
\usage{
import_operation_meta_basel(
  raw_dir_rhein = shiny_file("basel/data/operation/rhein"),
  raw_dir_wiese = shiny_file("basel/data/operation/wiese"),
  meta_online_path = shiny_file("basel/data/metadata/meta_online.csv"),
  meta_site_path = shiny_file("basel/data/metadata/meta_site.csv"),
  meta_parameter_path = shiny_file("basel/data/metadata/meta_parameter.csv")
)
}
\arguments{
\item{raw_dir_rhein}{Define directory for site "rhein" with raw data in
EXCEL spreadsheet format (.xlsx) to be imported (default:
sema.pilot:::shiny_file("basel/data/operation/rhein"))}

\item{raw_dir_wiese}{Define directory for site "rhein" with raw data in
EXCEL spreadsheet format (.xlsx) to be imported (default:
sema.pilot:::shiny_file("basel/data/operation/wiese"))}

\item{meta_online_path}{path to file containing metadata for online data
(default: sema.pilot:::shiny_file("basel/data/metadata/meta_online.csv"))}

\item{meta_site_path}{Define path of "meta_site.csv" to be imported
(default: sema.pilot:::shiny_file("basel/data/metadata/meta_site.csv"))}

\item{meta_parameter_path}{Define path of "meta_parameter.csv" to be imported
(default: sema.pilot:::shiny_file("basel/data/metadata/meta_parameter.csv"))}
}
\value{
returns data frame with imported raw operational data with metadata
for both sites (i.e."rhein" and "wiese")

data.frame with operational data for Basel sites including metadata
}
\description{
Imports operational data for Basel (with metadata for
both sites at once, i.e. "rhein" and "wiese")
}
