% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mbr4.R
\name{read_mbr4}
\alias{read_mbr4}
\title{Read MBR4.0 data from Martin Systems}
\usage{
read_mbr4(
  url = Sys.getenv("MBR4.0_URL"),
  target_path = file.path(tempdir(), "mbr4.tsv"),
  dbg = FALSE,
  ...
)
}
\arguments{
\item{url}{url to download site (default: Sys.getenv("MBR4.0_URL"), 
please use run \link[usethis]{edit_r_environ} and define a 
key value pair "MBR4.0_URL" = "download-url-martin-systems") so 
that this function works automatically}

\item{target_path}{to download data (default: file.path(tempdir(), "mbr4.tsv"))}

\item{dbg}{print debug messages (default: FALSE)}

\item{...}{additional arguments passed to  \link[readr]{read_tsv}}
}
\value{
tibble with imported MBR4.0 data
}
\description{
Read MBR4.0 data from Martin Systems
}
\examples{
mbr4_data <- read_mbr4()
str(mbr4_data)
}
