% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mbr4.R
\name{read_mbr4_archived}
\alias{read_mbr4_archived}
\title{Read MBR4.0 archived data from Nextcloud}
\usage{
read_mbr4_archived(
  file = "MBR_export_",
  dir = "projects/MBR4.0/Exchange/Rohdaten/Online_export",
  target_dir = tempdir(),
  locale = readr::locale(tz = "CET", decimal_mark = ",", grouping_mark = "."),
  col_types = readr::cols(.default = readr::col_double(), zustand =
    readr::col_character(), meldungen = readr::col_character(), Zeitstempel =
    readr::col_datetime(format = "\%d.\%m.\%Y \%H:\%M")),
  url = Sys.getenv("NEXTCLOUD_URL"),
  user = Sys.getenv("NEXTCLOUD_USER"),
  pw = Sys.getenv("NEXTCLOUD_USER"),
  dbg = FALSE,
  ...
)
}
\arguments{
\item{file}{file name/pattern of tsv file  (default: "MBR_export_")}

\item{dir}{directory on Nextcloud containing file (default: 
"projects/MBR4.0/Exchange/Rohdaten/Online_export")}

\item{target_dir}{directory to download data (default: tempdir())}

\item{locale}{locale (default: \code{\link[readr]{locale}}(tz = "CET",
decimal_mark = ",", grouping_mark = "."))}

\item{col_types}{col_types (default: \code{\link[readr]{cols}}(
.default = \code{\link[readr]{col_double}},  
 zustand  = \code{\link[readr]{col_character}},
 meldungen  = \code{\link[readr]{col_character}},
 Zeitstempel = \code{\link[readr]{col_datetime}}(format = "%d.%m.%Y %H:%M")
 )}

\item{url}{url of Nextcloud (default: Sys.getenv("NEXTCLOUD_URL"))}

\item{user}{username of Nextcloud  (default: Sys.getenv("NEXTCLOUD_USER"))}

\item{pw}{password of Nextcloud  (default: Sys.getenv("NEXTCLOUD_USER"))}

\item{dbg}{print debug messages (default: FALSE)}

\item{...}{additional arguments passed to  \link[readr]{read_tsv}}
}
\value{
tibble with imported archived MBR4.0 xlsx data from Nextcloud
}
\description{
Read MBR4.0 archived data from Nextcloud
}
\examples{
if(check_env_nextcloud()) {
mbr4_data_archived <- read_mbr4_archived()
str(mbr4_data_archived)
}
}
