% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mbr4.R
\name{read_mbr4_latest}
\alias{read_mbr4_latest}
\title{Read MBR4.0 data from Martin Systems Webportal (As "tsv")}
\usage{
read_mbr4_latest(
  url = Sys.getenv("MBR40_URL"),
  target_dir = tempdir(),
  locale = readr::locale(tz = "CET", decimal_mark = ".", grouping_mark = ","),
  col_types = readr::cols(.default = readr::col_double(), zustand =
    readr::col_character(), meldungen = readr::col_character(), Zeitstempel =
    readr::col_datetime(format = "\%Y-\%m-\%d \%H:\%M:\%S")),
  dbg = FALSE,
  ...
)
}
\arguments{
\item{url}{url to download site (default: Sys.getenv("MBR4.0_URL"),
please use run \link[usethis]{edit_r_environ} and define a
key value pair "MBR40_URL" = "download-url-martin-systems") so
that this function works automatically}

\item{target_dir}{directory to download data (default: tempdir())}

\item{locale}{locale (default: \code{\link[readr]{locale}}(tz = "CET",
decimal_mark = ".", grouping_mark = ","))}

\item{col_types}{col_types (default: \code{\link[readr]{cols}}(
.default = \code{\link[readr]{col_double}},  
 zustand  = \code{\link[readr]{col_character}},
 meldungen  = \code{\link[readr]{col_character}},
 Zeitstempel = \code{\link[readr]{col_datetime}}(format = "%Y-%m-%d %H:%M:%S")
 )}

\item{dbg}{print debug messages (default: FALSE)}

\item{...}{additional arguments passed to  \link[readr]{read_tsv}}
}
\value{
tibble with imported MBR4.0 tsv data (~ last four weeks)
}
\description{
Read MBR4.0 data from Martin Systems Webportal (As "tsv")
}
\examples{
url_mbr40 <- Sys.getenv("MBR40_URL")
if (url_mbr40 != "") {
  mbr4_data_latest <- read_mbr4_latest(url = url_mbr40)
  str(mbr4_data_latest)
}
}
