% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mbr4.R
\name{read_mbr4_tsv}
\alias{read_mbr4_tsv}
\title{Read MBR4.0 tsv data}
\usage{
read_mbr4_tsv(
  path,
  locale = readr::locale(tz = "CET", decimal_mark = ",", grouping_mark = "."),
  col_types = readr::cols(.default = readr::col_double(), zustand =
    readr::col_character(), meldungen = readr::col_character(), Zeitstempel =
    readr::col_datetime(format = "\%d.\%m.\%Y \%H:\%M")),
  dbg = FALSE,
  ...
)
}
\arguments{
\item{path}{path to tsv file to be imported}

\item{locale}{locale (default: \code{\link[readr]{locale}}(tz = "CET",
decimal_mark = ",", grouping_mark = "."))}

\item{col_types}{col_types (default: \code{\link[readr]{cols}}(
.default = \code{\link[readr]{col_double}},  
 zustand  = \code{\link[readr]{col_character}},
 meldungen  = \code{\link[readr]{col_character}},
 Zeitstempel = \code{\link[readr]{col_datetime}}(format = "%d.%m.%Y %H:%M")
 )}

\item{dbg}{print debug messages (default: FALSE)}

\item{...}{additional arguments passed to  \link[readr]{read_tsv}}
}
\value{
Reads MBR4.0 tsv data
}
\description{
Read MBR4.0 tsv data
}
