% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_weintek.R
\name{read_weintek}
\alias{read_weintek}
\title{read_weintek}
\usage{
read_weintek(path, tz = "CET", dbg = TRUE)
}
\arguments{
\item{path}{path to Weintek file}

\item{tz}{time zone (default: CET)}

\item{dbg}{debug (default: TRUE)}
}
\value{
data frame with Weintek raw data
}
\description{
read_weintek
}
