% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_weintek.R
\name{read_weintek_batch}
\alias{read_weintek_batch}
\title{read_weintek_batch}
\usage{
read_weintek_batch(files, tz = "CET", dbg = TRUE)
}
\arguments{
\item{files}{path to Weintek files}

\item{tz}{time zone (default: CET)}

\item{dbg}{debug (default: TRUE)}
}
\value{
data frame with Weintek raw data
}
\description{
read_weintek_batch
}
