% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influxdb_ultimate.R
\name{write_to_influxdb_loop}
\alias{write_to_influxdb_loop}
\title{InfluxDB: write to InfluxDB in Loop}
\usage{
write_to_influxdb_loop(tsv_paths, paths, max_tsv_files = 5, batch_size = 5000)
}
\arguments{
\item{tsv_paths}{vector with tsv_paths with files to be imported by 
\code{\link{write_to_influxdb}} which relies on \code{\link{read_pentair_data}}}

\item{paths}{paths list with elements \code{raw_data_dir}, \code{site_code}, 
\code{influx_url}, \code{influx_token} and \code{influx_org}}

\item{max_tsv_files}{maximum number of tsv files to read at once (should be
limited due to high RAM demand), default: 5}

\item{batch_size}{number of data points that are written in one query (default:
5000)}
}
\value{
writes imported data to InfluxDB in Loop
}
\description{
wrapper function for \code{\link{write_to_influxdb}}
}
